# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: XLAIR API
  description: ""
  license:
    name: ""
  version: "1.0.0"
...
servers:
  - url: https://api.xlair.dev
paths:
  /users:
    post:
      tags:
        - station
      summary: ユーザー登録
      description: カードIDで筐体に新規ログインした時に呼ぶ
      security:
        - deviceAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/userRegister"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userData"
        "400":
          description: Bad request - Invalid input data
        "401":
          description: Unauthorized - Invalid device access token
        "409":
          description: Conflict - User with this card ID already exists
        "500":
          description: Internal server error
    get:
      tags:
        - app
        - station
      summary: カードIDからユーザー情報を取得
      description: カードIDで筐体に通常ログインした時に呼ぶ
      security:
        - deviceAuth: []
      parameters:
        - name: card
          in: query
          description: カードID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userData"
        "400":
          description: Bad request - Missing or invalid card ID
        "401":
          description: Unauthorized - Invalid device access token
        "404":
          description: Not found - User with specified card ID not found
        "500":
          description: Internal server error
  /users/{userId}:
    patch:
      tags:
        - station
      summary: ユーザー情報の更新
      description: ユーザーの表示名と公開設定を更新する
      security:
        - deviceAuth: []
      parameters:
        - name: userId
          in: path
          description: ユーザーのID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/userUpdate"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userData"
        "401":
          description: Unauthorized - Invalid access token
        "403":
          description: Forbidden - The authenticated principal cannot update this user
        "404":
          description: Not found - User not found
        "500":
          description: Internal server error
  /admin/db/synchronize:
    post:
      tags:
        - admin
      summary: DBの派生値を同期
      description: 現在のドメインルールに基づいて、DBに保存された派生値を同期する。値が変わるユーザーのみ更新する。
      security:
        - userAuth: []
      responses:
        "200":
          description: Recalculation completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  updatedUsers:
                    type: integer
                    format: int64
                  updatedRatings:
                    type: integer
                    format: int64
                required:
                  - updatedUsers
                  - updatedRatings
        "401":
          description: Unauthorized - Invalid user access token
        "403":
          description: Forbidden - The authenticated user is not an administrator
        "500":
          description: Internal server error
  /admin/musics:
    post:
      tags:
        - admin
      summary: 楽曲と譜面を登録
      security:
        - userAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/musicData"
                - type: object
                  properties:
                    sheets:
                      type: array
                      minItems: 3
                      maxItems: 3
                      items:
                        $ref: "#/components/schemas/sheetData"
                  required:
                    - sheets
      responses:
        "201":
          description: 登録された楽曲と譜面
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/musicWithSheets"
        "400":
          description: Bad request - Invalid music or sheet data
        "401":
          description: Unauthorized - Invalid access token
        "403":
          description: Forbidden - The authenticated user is not an administrator
        "500":
          description: Internal server error
    get:
      tags:
        - admin
      summary: 楽曲と譜面の一覧を取得
      description: 楽曲を登録日時とIDの昇順で取得する。cursorを指定すると続きから取得する。
      security:
        - userAuth: []
      parameters:
        - name: cursor
          in: query
          description: 前ページのレスポンスに含まれる次のページ用カーソル
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: 取得件数 (1〜100、既定値50)
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: 楽曲と譜面のページ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/adminMusicListResponse"
        "400":
          description: Bad request - Invalid cursor or limit
        "401":
          description: Unauthorized - Invalid user access token
        "403":
          description: Forbidden - The authenticated user is not an administrator
        "500":
          description: Internal server error
  /admin/musics/{musicId}:
    get:
      tags:
        - admin
      summary: 楽曲と譜面の詳細を取得
      security:
        - userAuth: []
      parameters:
        - name: musicId
          in: path
          description: 楽曲のID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 楽曲と譜面の詳細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/musicWithSheets"
        "401":
          description: Unauthorized - Invalid access token
        "403":
          description: Forbidden - The authenticated user is not an administrator
        "404":
          description: Not found - Music not found
        "500":
          description: Internal server error
    post:
      tags:
        - admin
      summary: 楽曲と譜面を更新
      security:
        - userAuth: []
      parameters:
        - name: musicId
          in: path
          description: 楽曲のID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/musicData"
                - type: object
                  properties:
                    sheets:
                      type: array
                      minItems: 3
                      maxItems: 3
                      items:
                        $ref: "#/components/schemas/sheet"
                  required:
                    - sheets
      responses:
        "200":
          description: 更新された楽曲と譜面
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/musicWithSheets"
        "400":
          description: Bad request - Invalid music or sheet data
        "401":
          description: Unauthorized - Invalid access token
        "403":
          description: Forbidden - The authenticated user is not an administrator
        "404":
          description: Not found - Music not found
        "500":
          description: Internal server error
  /admin/musics/{musicId}/jacket:
    post:
      tags:
        - admin
      summary: 楽曲のジャケット画像をアップロード
      security:
        - userAuth: []
      parameters:
        - name: musicId
          in: path
          description: 楽曲のID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
          image/png:
            schema:
              type: string
              format: binary
          image/webp:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: ジャケット画像が更新された楽曲と譜面
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/musicWithSheets"
        "400":
          description: Bad request - Invalid jacket image
        "401":
          description: Unauthorized - Invalid access token
        "403":
          description: Forbidden - The authenticated user is not an administrator
        "404":
          description: Not found - Music not found
        "500":
          description: Internal server error
    delete:
      tags:
        - admin
      summary: 楽曲のジャケット画像を削除
      security:
        - userAuth: []
      parameters:
        - name: musicId
          in: path
          description: 楽曲のID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: ジャケット画像が削除された楽曲と譜面
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/musicWithSheets"
        "400":
          description: Bad request - Invalid music ID
        "401":
          description: Unauthorized - Invalid access token
        "403":
          description: Forbidden - The authenticated user is not an administrator
        "404":
          description: Not found - Music not found
        "500":
          description: Internal server error
  /admin/musics/{musicId}/audio:
    post:
      tags: [admin]
      summary: 楽曲の音源をアップロード
      security:
        - userAuth: []
      parameters:
        - $ref: "#/components/parameters/musicId"
      requestBody:
        required: true
        content:
          audio/wav:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: 音源が更新された楽曲と譜面
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/musicWithSheets"
        "400": { description: Bad request - Invalid WAV file }
        "401": { description: Unauthorized - Invalid access token }
        "403": { description: Forbidden - The authenticated user is not an administrator }
        "404": { description: Not found - Music not found }
        "500": { description: Internal server error }
    delete:
      tags: [admin]
      summary: 楽曲の音源を削除
      security:
        - userAuth: []
      parameters:
        - $ref: "#/components/parameters/musicId"
      responses:
        "200":
          description: 音源が削除された楽曲と譜面
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/musicWithSheets"
        "401": { description: Unauthorized - Invalid access token }
        "403": { description: Forbidden - The authenticated user is not an administrator }
        "404": { description: Not found - Music not found }
        "500": { description: Internal server error }
  /admin/musics/{musicId}/audio/{fileName}:
    get:
      tags: [admin]
      summary: 楽曲の音源を取得
      security:
        - userAuth: []
      parameters:
        - $ref: "#/components/parameters/musicId"
        - $ref: "#/components/parameters/assetFileName"
      responses:
        "200":
          description: WAV 音源
          content:
            audio/wav:
              schema: { type: string, format: binary }
        "401": { description: Unauthorized - Invalid access token }
        "403": { description: Forbidden - The authenticated user is not an administrator }
        "404": { description: Not found - Music or audio not found }
        "500": { description: Internal server error }
  /admin/sheets/{sheetId}/chart:
    post:
      tags: [admin]
      summary: 譜面ファイルをアップロード
      security:
        - userAuth: []
      parameters:
        - name: sheetId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        "200":
          description: 譜面ファイルが更新された楽曲と譜面
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/musicWithSheets"
        "400": { description: Bad request - Invalid SUS file name }
        "401": { description: Unauthorized - Invalid access token }
        "403": { description: Forbidden - The authenticated user is not an administrator }
        "404": { description: Not found - Sheet not found }
        "500": { description: Internal server error }
    delete:
      tags: [admin]
      summary: 譜面ファイルを削除
      security:
        - userAuth: []
      parameters:
        - name: sheetId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: 譜面ファイルが削除された楽曲と譜面
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/musicWithSheets"
        "401": { description: Unauthorized - Invalid access token }
        "403": { description: Forbidden - The authenticated user is not an administrator }
        "404": { description: Not found - Sheet not found }
        "500": { description: Internal server error }
  /admin/sheets/{sheetId}/chart/{fileName}:
    get:
      tags: [admin]
      summary: 譜面ファイルを取得
      security:
        - userAuth: []
      parameters:
        - name: sheetId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/assetFileName"
      responses:
        "200":
          description: SUS 譜面ファイル
          content:
            text/plain:
              schema: { type: string, format: binary }
        "401": { description: Unauthorized - Invalid access token }
        "403": { description: Forbidden - The authenticated user is not an administrator }
        "404": { description: Not found - Sheet or chart not found }
        "500": { description: Internal server error }
  /musics/{musicId}/jacket/{fileName}:
    get:
      tags: [asset]
      summary: ジャケット画像を取得
      parameters:
        - $ref: "#/components/parameters/musicId"
        - $ref: "#/components/parameters/assetFileName"
      responses:
        "200":
          description: ジャケット画像
          content:
            image/png:
              schema: { type: string, format: binary }
        "404": { description: Not found - Jacket not found }
  /musics/{musicId}/audio/{fileName}:
    get:
      tags: [asset]
      summary: 楽曲の音源を取得
      security:
        - deviceAuth: []
      parameters:
        - $ref: "#/components/parameters/musicId"
        - $ref: "#/components/parameters/assetFileName"
      responses:
        "200":
          description: WAV 音源
          content:
            audio/wav:
              schema: { type: string, format: binary }
        "401": { description: Unauthorized - Invalid access token }
        "404": { description: Not found - Audio not found }
  /sheets/{sheetId}/chart/{fileName}:
    get:
      tags: [asset]
      summary: 譜面ファイルを取得
      security:
        - deviceAuth: []
      parameters:
        - name: sheetId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/assetFileName"
      responses:
        "200":
          description: SUS 譜面ファイル
          content:
            text/plain:
              schema: { type: string, format: binary }
        "401": { description: Unauthorized - Invalid access token }
        "404": { description: Not found - Chart not found }
  /users/{userId}/credits/increment:
    post:
      tags:
        - app
      summary: ユーザーのクレジット数を更新
      description: ユーザーのクレジット数を1増やす。ゲーム開始時に呼ぶ
      security:
        - deviceAuth: []
      parameters:
        - name: userId
          in: path
          description: ユーザーのID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  credits:
                    type: integer
                    description: 更新後のクレジット数
        "401":
          description: Unauthorized - Invalid device access token
        "404":
          description: Not found - User not found
        "500":
          description: Internal server error
  /users/{userId}/records:
    get:
      tags:
        - app
      summary: ユーザーのプレイデータを取得
      description: ゲーム開始時にユーザーのプレイデータを一括取得する
      security:
        - deviceAuth: []
      parameters:
        - name: userId
          in: path
          description: ユーザーのID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/record"
        "401":
          description: Unauthorized - Invalid device access token
        "404":
          description: Not found - User not found
        "500":
          description: Internal server error
    post:
      tags:
        - app
      summary: ユーザーのプレイデータを送信
      description: スコアの更新の有無に関わらず、ゲーム終了時にユーザーのプレイデータを送信する
      security:
        - deviceAuth: []
      parameters:
        - name: userId
          in: path
          description: ユーザーのID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/recordRequest"
      responses:
        "201":
          description: Records created/updated successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/record"
        "400":
          description: Bad request - Invalid input data
        "401":
          description: Unauthorized - Invalid device access token
        "404":
          description: Not found - User or sheet not found
        "500":
          description: Internal server error
  /users/{userId}/options:
    get:
      tags:
        - app
      summary: ユーザーのプレイオプションを取得
      description: ゲーム開始時にユーザーのプレイオプションを取得する
      security:
        - deviceAuth: []
      parameters:
        - name: userId
          in: path
          description: ユーザーのID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userPlayOption"
        "401":
          description: Unauthorized - Invalid device access token
        "404":
          description: Not found - User not found
        "500":
          description: Internal server error
    post:
      tags:
        - app
      summary: ユーザーのプレイオプションを更新
      description: ゲーム終了時にユーザーのプレイオプションを保存する
      security:
        - deviceAuth: []
      parameters:
        - name: userId
          in: path
          description: ユーザーのID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/userPlayOption"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userPlayOption"
        "401":
          description: Unauthorized - Invalid device access token
        "404":
          description: Not found - User not found
        "500":
          description: Internal server error
  /sync:
    get:
      tags:
        - app
      summary: 全曲・全譜面のメタデータを取得
      description: 同期用。筐体が起動した最初の一回しか実行しない
      security:
        - deviceAuth: []
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/musicWithSheets"
        "401":
          description: Unauthorized - Invalid device access token
        "500":
          description: Internal server error
  /rankings/sheets/{sheetId}:
    get:
      tags:
        - web
      summary: 指定した譜面のハイスコアランキングを取得
      description: 公開ユーザーを対象に、指定した譜面のハイスコア上位20件を取得する
      parameters:
        - name: sheetId
          in: path
          description: 譜面のID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/sheetScoreRankingResponse"
        "404":
          description: Not found - Sheet not found
        "500":
          description: Internal server error
  /rankings/total-score:
    get:
      tags:
        - web
      summary: 全譜面のハイスコア合計ランキングを取得
      description: 公開ユーザーを対象に、保有する全譜面のハイスコア合計上位20件を取得する
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/totalScoreRankingResponse"
        "500":
          description: Internal server error
  /rankings/rating:
    get:
      tags:
        - web
      summary: レーティングのランキングを取得
      description: 公開ユーザーを対象に、レーティングの上位20件を取得する
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ratingRankingResponse"
        "500":
          description: Internal server error
  /rankings/xp:
    get:
      tags:
        - web
      summary: 経験値のランキングを取得
      description: 公開ユーザーを対象に、XP の上位20件を取得する
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/xpRankingResponse"
        "500":
          description: Internal server error
  /statistics/summary:
    get:
      tags:
        - web
      summary: システム全体の統計情報を取得
      description: 全ユーザーとレコードの集計値を参照する
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/globalStatistics"
        "500":
          description: Internal server error
...
components:
...
  securitySchemes:
    userAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Auth0 によるユーザー認証トークン。
        現時点では管理者・運用者用 dashboard で利用する。
        一般ユーザーは Auth0 アカウントを作成せず、筐体では card ID で識別する。
    deviceAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Auth0 Machine-to-Machine Application による筐体認証トークン。
        XLAIR が認証主体を筐体として分類するための大分類を含む。
  schemas:
    userRegister:
      type: object
      properties:
        card:
          type: string
          description: カードID
        displayName:
          type: string
          description: ユーザーの表示名
        isPublic:
          type: boolean
          description: ユーザー情報を xlair.dev 上に公開するかどうかのフラグ
      required:
        - card
        - displayName
        - isPublic
    userUpdate:
      type: object
      properties:
        displayName:
          type: string
          description: ユーザーの表示名
        isPublic:
          type: boolean
          description: ユーザー情報を xlair.dev 上に公開するかどうかのフラグ
      required:
        - displayName
        - isPublic
    userData:
      type: object
      properties:
        id:
          type: string
          description: ユーザーのID
        card:
          type: string
          description: Felica の IDm
        displayName:
          type: string
          description: ユーザーの表示名
        rating:
          type: integer
          description: ユーザーのレーティング
        xp:
          type: integer
          description: ユーザーの経験値
        credits:
          type: integer
          description: ユーザーの総クレジット数
        isPublic:
          type: boolean
          description: ユーザーの情報を xlair.dev 上に公開するかどうかのフラグ
        isAdmin:
          type: boolean
          description: ユーザーが管理者かどうかを表すフラグ
        createdAt:
          type: string
          format: date-time
          description: ユーザーの登録日時
      required:
        - id
        - card
        - displayName
        - rating
        - xp
        - credits
        - isPublic
        - isAdmin
        - createdAt
...
    musicData:
      type: object
      properties:
        title:
          type: string
          description: タイトル
        artist:
          type: string
          description: アーティスト名
        bpm:
          type: number
          format: float
          description: BPM
        genre:
          type: string
          enum:
            - ORIGINAL
            - EXTERNAL
            - OTHER
          description: ジャンル
        registrationDate:
          type: string
          format: date-time
          description: 運用上の楽曲登録日
        isTest:
          type: boolean
          description: テスト楽曲かどうか
      required:
        - title
        - artist
        - bpm
        - genre
        - registrationDate
        - isTest
...
    sheetData:
      type: object
      properties:
        difficulty:
          type: string
          enum:
            - basic
            - advanced
            - master
          description: 難易度の分類
        level:
          type: number
          description: レベル
        notesDesigner:
          type: string
          description: 譜面のノーツデザイナー
      required:
        - difficulty
        - level
        - notesDesigner
...
    musicWithSheets:
      type: object
      properties:
        music:
          $ref: "#/components/schemas/music"
        sheets:
          type: array
          items:
            $ref: "#/components/schemas/sheetWithMusicId"
      required:
        - music
        - sheets
    adminMusicListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/musicWithSheets"
        nextCursor:
          type: string
          nullable: true
      required:
        - items
        - nextCursor
...